Separate cat generation from canvas buffer generation
